+++ src/neo/SmartContract/Native/Designate/Role.cs
@@ -1,0 +1,9 @@
+
+namespace Neo.SmartContract.Native.Designate
+{
+    public enum Role : byte
+    {
+        StateValidator = 4,
+        Oracle = 8
+    }
+}
